separate objutil from util
